Add the -r option (reverse task list) and -m option (max nbr tasks listed).
These options can be useful if you have a small command
window, and more tasks to display than can fit without
scrolling out of view.

The max task count (-m) option restricts the task list
to just some integer number of tasks (e.g. -m 20). This
option works with list, listall, and listpri, but not
with listfile.

The reverse (-r) option mitigates the scrolling-out-of-view
issue by listing the tasks in reverse sequence, so the most
important tasks are listed last.  Like the -m option, it
works with list, listall, and listpri, but not with listfile.