Made version and short usage actions backward compatible.
Philippe Teuwen pointed out that using the new actions "version"
and "usage-short" would not be backward compatible, because
addons that don't support these would interpret them as
command arguments. So I've changed things to use "version"
and "short" as arguments to the usage action. This should not
break actions that do not expect the new arguments because
they will just to the 'usual' thing when they see the action
'usage'. (Thanks Phil!)