Added the ellipsis support. But it doesn't quite work.
The problem is that the -+@P processing happens after the
TODO_SORT_COMMAND is invoked, so we don't know the line length.
And we have to do the sort before adding the line number and colors.
Only solution I see is to add a TODO_TXT_FINALFILTER that
is processed after the colors are added.