This change is intended to allow the config file to use simple key=value assignments, with short functional key names, and without worrying about the use of 'export'.
The change also separates the namespace of the variables
used by the script from the namespace of the variables
in the config file; the internally used script variables
begin with the 'TODOTXT_' prefix.

Furthermore, the changes are intended to normalize the
precedence sequence by which the script variables are
set. This change allow addons to override the script
variables without getting stepped in the second invocation
of the script (a second invocation via the 'command' action
is a standard pattern for implementing addons). The order
of precedence is:

  1. command line option, else
  2. environment variables, else
  3. configuration file, else
  4. hard-coded defaults

Finally, additional config file key-value pairs were
defined to match some of the option flags.

All the tests passed, though some had to be modified
to accomodate modified script variable names and to
avoid the use of 'export' in the test's config file.

Note: This commit is a port of Philippe Teuwen's original
work (see his repository branch 'new_export'), but based
on the latest production 'master' branch.

Note2: To implement the partition of the namespace
between configuration variables and script variables,
some script variable names have changed; this may
have an impact to some addons.